discord_bot_project/requirements.txt

discord.py
discord-py-slash-command
aiohttp==3.11.11
requests==2.32.3
flask==3.1.0
psycopg==3.2.3
pytz==2023.3
python-dotenv==0.21.1
pip==24.3.1
interaction==1.3
